Moved legacy tests to separate package
