+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/legacy/AbstractPostgresSourceAcceptanceTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres
+package io.airbyte.integrations.source.postgres.legacy
 
 import io.airbyte.cdk.test.fixtures.legacy.SourceAcceptanceTest
 import io.airbyte.cdk.test.fixtures.legacy.SshHelpers
