+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/legacy/AbstractPostgresSourceDatatypeTest.kt
@@ -2,11 +2,12 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres
+package io.airbyte.integrations.source.postgres.legacy
 
 import io.airbyte.cdk.test.fixtures.legacy.AbstractSourceDatabaseTypeTest
 import io.airbyte.cdk.test.fixtures.legacy.TestDataHolder
 import io.airbyte.cdk.test.fixtures.legacy.TestDestinationEnv
+import io.airbyte.integrations.source.postgres.PostgresTestDatabase
 import io.airbyte.protocol.models.JsonSchemaPrimitiveUtil
 import io.airbyte.protocol.models.JsonSchemaType
 import java.util.Locale
