+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/legacy/PostgresSourceAcceptanceLegacyCtidTest.kt
@@ -2,7 +2,9 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres
+package io.airbyte.integrations.source.postgres.legacy
+
+import io.airbyte.integrations.source.postgres.PostgresTestDatabase
 
 class PostgresSourceAcceptanceLegacyCtidTest : PostgresSourceAcceptanceTest() {
     override val serverImage: PostgresTestDatabase.BaseImage
