+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/legacy/PostgresSourceDatatypeTest.kt
@@ -2,18 +2,19 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres
+package io.airbyte.integrations.source.postgres.legacy
 
 import com.fasterxml.jackson.databind.JsonNode
 import io.airbyte.cdk.test.fixtures.legacy.Database
 import io.airbyte.cdk.test.fixtures.legacy.JdbcUtils.MODE_KEY
+import io.airbyte.integrations.source.postgres.PostgresTestDatabase
 import java.sql.SQLException
 
 class PostgresSourceDatatypeTest : AbstractPostgresSourceDatatypeTest() {
     @Throws(SQLException::class)
     protected override fun setupDatabase(): Database {
         testdb =
-            PostgresTestDatabase.`in`(
+            PostgresTestDatabase.Companion.`in`(
                     PostgresTestDatabase.BaseImage.POSTGRES_17,
                     PostgresTestDatabase.ContainerModifier.CONF
                 )
